airflow/requirements.txt
pandas 
requests 
mpsycopg2 
sqlalchemy